[psqldef] Treat 'timestamptz' the same as 'timestamp with time zone' (#545)
When parsing a column with type 'timestamptz', we don't go through the code path
which normalizes 'timestamptz' to 'timestamp with time zone'. This causes
psqldef to try to change the type of a timestamptz column.

Use the normalization codepath by default, but include some compatibility
checks to avoid changing existing tests.