Add --dump-concurrency option
This new command-line option controls the concurrency
level of queries during DB schema dump process.

It speeds up the DB schema dump process when enabled.

This option is particularly effective when the database
is located remotely and contains many tables, as network
latency becomes the dominant factor in the total duration
of the dump process.