Fix "TABLELOCK" to "TABLOCK" for MSSQL.
The keyword `TABLELOCK` defined in the parser appears to be a mistake, as there is no known RDBMS that uses `TABLELOCK` as a lock hint keyword.
This likely originated as a typographical error. All other lock hints defined in this section (e.g., `ROWLOCK`, `PAGLOCK`, `NOLOCK`) correctly follow SQL Server’s syntax, making this inconsistency stand out.
If this part of the parser is intended to target SQL Server, the correct table-level locking hint should be `TABLOCK`.
https://learn.microsoft.com/en-us/sql/t-sql/queries/hints-transact-sql-table?view=sql-server-ver17#syntax